backend/pyproject.toml
[project]
name = "remind-invoice-backend"
version = "0.1.0"
description = "SaaS for freelancers to create invoices and automate payment reminders"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.1",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "email-validator>=2.1.0",
    "pydantic-settings>=2.1.0",
    "slowapi>=0.1.9",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "sendgrid>=6.11.0",
    "weasyprint>=60.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]
